Views improved in Patient details, fab button removed from Agent Patient view, API 19 crash fixed
