+++ app/src/main/res/drawable/ic_menu_camera.xml
