+++ app/src/main/res/drawable/ic_menu_gallery.xml
