+++ app/src/main/res/drawable/ic_menu_manage.xml
