+++ app/src/main/res/drawable/ic_prescription_pill.png
