+++ app/src/main/res/drawable/ic_thermometer.png
