+++ app/src/main/res/layout/activity_patient_dr.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -12,8 +11,8 @@
         android:id="@+id/appbarPatientDr"
         android:layout_width="match_parent"
         android:layout_height="320dp"
-        android:fitsSystemWindows="true"
         android:background="@drawable/shape_gradient"
+        android:fitsSystemWindows="true"
         android:theme="@style/AppTheme.AppBarOverlay">
 
         <android.support.design.widget.CollapsingToolbarLayout
@@ -27,8 +26,8 @@
                 android:id="@+id/tbPatientDr"
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize"
-                app:layout_collapseMode="pin"
                 android:layout_gravity="top"
+                app:layout_collapseMode="parallax"
                 app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
                 app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
 
@@ -39,10 +38,12 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_gravity="bottom"
-                app:tabTextColor="@color/patientDrTab"
+                app:tabGravity="fill"
                 app:tabMode="fixed"
-                app:tabGravity="fill" />
+                app:tabTextColor="@color/patientDrTab" />
+
         </android.support.design.widget.CollapsingToolbarLayout>
+
     </android.support.design.widget.AppBarLayout>
 
     <android.support.v4.view.ViewPager
@@ -53,13 +54,13 @@
 
 
     <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fabAdd"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_gravity="bottom|right"
+        android:layout_margin="16dp"
         android:src="@drawable/ic_prescription_pill"
         app:fabSize="normal"
-        android:id="@+id/fabAdd"
-        android:layout_margin="16dp"
-        android:layout_gravity="bottom|right"
         app:layout_anchor="@id/containerPatientDr"
-        app:layout_anchorGravity="bottom|right|end"/>
+        app:layout_anchorGravity="bottom|right|end" />
 </android.support.design.widget.CoordinatorLayout>
