+++ app/src/main/res/menu/main_activity2.xml
@@ -2,8 +2,8 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/action_settings"
+        android:id="@+id/action_signout"
         android:orderInCategory="100"
-        android:title="@string/action_settings"
+        android:title="@string/action_signout"
         app:showAsAction="never" />
 </menu>
